update to pytorch1.0.

+++ extensions/make.sh
@@ -9,25 +9,28 @@ echo "Building cocoapi..."
 cd apis/cocoapi/PythonAPI
 ${PYTHON} setup.py install
 
-cd -
 echo "Building roi align op..."
-cd ./roi_align
+cd ops/roi_align
 if [ -d "build" ]; then
     rm -r build
 fi
-${PYTHON} setup.py build_ext --inplace
-rm -r build
+$PYTHON setup.py build_ext --inplace
 
 echo "Building roi pool op..."
 cd ../roi_pool
 if [ -d "build" ]; then
     rm -r build
 fi
-${PYTHON} setup.py build_ext --inplace
-rm -r build
+$PYTHON setup.py build_ext --inplace
 
 echo "Building nms op..."
-cd ../nms/src
+cd ../nms
 make clean
 make PYTHON=${PYTHON}
-rm -r build
+
+echo "Building dcn..."
+cd ../dcn
+if [ -d "build" ]; then
+    rm -r build
+fi
+$PYTHON setup.py build_ext --inplace
